condense files and cleanup
