+++ include/code_80958E8.h
@@ -11,6 +11,15 @@ typedef struct subStruct_203B490
     /* 0x8 */ s32 checksum; // really dumb checksum that just adds all the fields of WonderMail struct
 } subStruct_203B490;
 
+
+// size: 0x8
+struct unkStruct_8096AF8
+{
+    bool8 unk0;
+    /* 0x2 */ u16 clientSpecies;
+    /* 0x4 */ u16 targetSpecies;
+};
+
 // size: 0x32C?
 typedef struct unkStruct_203B490
 {
@@ -35,5 +44,27 @@ s32 CountFilledMailboxSlots(void);
 WonderMail *GetMailboxSlotInfo(u8 index);
 bool8 IsMailSlotEmpty(u8 index);
 void ResetMailboxSlot(u8 index);
+bool8 AcceptJob(WonderMail *mail);
+s32 CountJobsinDungeon(u8 dungeon);
+WonderMail *GetJobSlotInfo(u8 index);
+s32 GetNumAcceptedJobs(void);
+bool8 IsJobSlotEmpty(u8 index);
+bool8 IsMailinJobSlot(WonderMail *mail);
+void ResetJobSlot(u8 index);
+bool8 sub_809693C(WonderMail *mail);
+void sub_8096AF8(struct unkStruct_8096AF8 *, u8 slotIndex, u8 dungeon);
+bool8 sub_8096C08(u8 *jobIndex);
+WonderMail *GetPelipperBoardSlotInfo(u8);
+bool8 IsPelipperBoardSlotEmpty(u8);
+void ResetPelipperBoardSlot(u8);
+void sub_80965F4(void);
+void sub_8096C80(void);
+void ReceivePKMNNews(u8 index);
+bool8 CheckPKMNNewsSlot(u8 index);
+void SortJobSlots(void);
+u8 *sub_8096DD8(void);
+u8 *sub_8096DE8(void);
+u8 sub_8096E2C(void);
+void sub_8096EEC(WonderMail *mail);
 
 #endif // GUARD_CODE_80958E8_H
