+++ src/code_801D014.c
@@ -8,7 +8,7 @@
 #include "code_80227B8.h"
 #include "code_8023868.h"
 #include "code_8024458.h"
-#include "code_8096AF8.h"
+#include "code_80958E8.h"
 #include "common_strings.h"
 #include "constants/friend_area.h"
 #include "event_flag.h"
