+++ src/code_802DB28.c
@@ -4,8 +4,7 @@
 #include "code_802DB28.h"
 #include "code_802F204.h"
 #include "code_803C1D0.h"
-#include "code_80958E8_1.h"
-#include "code_8096AF8.h"
+#include "code_80958E8.h"
 #include "exclusive_pokemon.h"
 #include "input.h"
 #include "items.h"
