+++ src/code_80972F4.c
@@ -1,7 +1,4 @@
 #include "global.h"
-#include "constants/dungeon.h"
-#include "pokemon.h"
-#include "code_80958E8.h"
 #include "dungeon.h"
 #include "exclusive_pokemon.h"
 #include "event_flag.h"
@@ -14,272 +11,9 @@ ALIGNED(4) const u8 gDummyScenarioText[] = _("{COLOR_1 YELLOW_3}Scenario try dum
 ALIGNED(4) const u8 gBlankMission[] = _("{COLOR_1 RED_2}???{END_COLOR_TEXT_1}");
 
 extern void sub_800199C(u8, u8, u16, s32);
-extern void SaveDungeonLocation(struct unkStruct_8094924*, DungeonLocation*);
-extern void RestoreDungeonLocation(struct unkStruct_8094924*, DungeonLocation*);
 extern u32 sub_8001784(u32, u32, u16);
 extern void GeneratePelipperJobs(void);
 extern void sub_80018D8(u8, u8, u32);
-
-void SaveWonderMail(struct unkStruct_8094924 *a, WonderMail *b);
-void RestoreWonderMail(struct unkStruct_8094924 *a, WonderMail *b);
-s32 CalculateMailChecksum(WonderMail *mail);
-
-void SortJobSlots(void)
-{
-    s32 r1;
-    s32 r6;
-    WonderMail job;
-    
-    for(r1 = 0; r1 < MAX_ACCEPTED_JOBS - 1; r1++)
-    {
-         for(r6 = r1 + 1; r6 < MAX_ACCEPTED_JOBS; r6++)
-        {
-                if(gUnknown_203B490->jobSlots[r6].mailType != 0)
-                {
-                    if((gUnknown_203B490->jobSlots[r1].unk4.dungeon.id > gUnknown_203B490->jobSlots[r6].unk4.dungeon.id) || 
-                        ((gUnknown_203B490->jobSlots[r1].unk4.dungeon.id == gUnknown_203B490->jobSlots[r6].unk4.dungeon.id) && 
-                        (gUnknown_203B490->jobSlots[r1].unk4.dungeon.floor > gUnknown_203B490->jobSlots[r6].unk4.dungeon.floor)))
-                    {
-                        job = gUnknown_203B490->jobSlots[r1];
-                        gUnknown_203B490->jobSlots[r1] = gUnknown_203B490->jobSlots[r6];
-                        gUnknown_203B490->jobSlots[r6] =  job;
-                    }
-                }
-        }
-    }
-}
-
-u8 *sub_8096DD8(void)
-{
-    return gUnknown_203B490->unk190;
-}
-
-u8 *sub_8096DE8(void)
-{
-    return gUnknown_203B490->unk1B8;
-}
-
-void ReceivePKMNNews(u8 index)
-{
-    gUnknown_203B490->PKMNNewsReceived[index] = TRUE;
-}
-
-bool8 CheckPKMNNewsSlot(u8 index)
-{
-    return gUnknown_203B490->PKMNNewsReceived[index];
-}
-
-u8 sub_8096E2C(void)
-{
-    s32 index;
-    s32 floor = 0;
-    for(index = 0; index < NUM_MAILBOX_SLOTS; index++)
-    {
-        if(gUnknown_203B490->mailboxSlots[index].mailType == 1)
-        {
-             if(floor <= gUnknown_203B490->mailboxSlots[index].unk4.dungeon.floor)
-                 if( gUnknown_203B490->mailboxSlots[index].unk4.dungeon.floor < 0x32) 
-                    floor =  gUnknown_203B490->mailboxSlots[index].unk4.dungeon.floor + 1;
-        }
-    }
-
-    for(index = floor; index < 0x32; index++)
-    {
-        if(!gUnknown_203B490->PKMNNewsReceived[index]) return index;
-    }
-    return 0x38;
-}
-
-bool8 sub_8096E80(u8 floor)
-{
-    s32 index;
-    for(index = 0; index < NUM_MAILBOX_SLOTS; index++)
-    {
-        if(gUnknown_203B490->mailboxSlots[index].mailType == 1)
-        {
-             if(floor == gUnknown_203B490->mailboxSlots[index].unk4.dungeon.floor)
-                return TRUE;
-        }
-    }
-    return FALSE;
-}
-
-s32 CalculateMailChecksum(WonderMail *mail)
-{
-    s32 sum;
-
-    sum = (mail->unk2 + mail->missionType);
-    
-    sum += mail->unk4.dungeon.id;
-    sum += mail->unk4.dungeon.floor;
-
-    sum += mail->unk4.seed << 0x8;
-
-    sum += mail->clientSpecies << 0xC;
-
-    sum += mail->targetSpecies << 0x10;
-
-    sum += mail->targetItem << 0x18;
-
-    sum += mail->rewardType;
-
-    sum += mail->itemReward;
-
-    sum += mail->friendAreaReward;
-
-    return sum;
-}
-
-void sub_8096EEC(WonderMail *mail)
-{
-    s32 index;
-
-    for(index = 15; index > 0; index--)
-    {
-         gUnknown_203B490->unk230[index] =  gUnknown_203B490->unk230[index - 1];
-    }
-    gUnknown_203B490->unk230[0].sub = mail->unk4;
-    gUnknown_203B490->unk230[0].checksum = CalculateMailChecksum(mail);
-}
-
-bool8 sub_8096F50(WonderMail *mail)
-{
-    s32 index;
-    s32 checksum;
-    subStruct_203B490 *temp;
-
-    checksum = CalculateMailChecksum(mail);
-
-    for (index = 0; index < 0x10; index++) {
-        temp  = &gUnknown_203B490->unk230[index];
-        if (temp->sub.dungeon.id == mail->unk4.dungeon.id)
-            if (temp->sub.dungeon.floor == mail->unk4.dungeon.floor)
-                if (temp->sub.seed == mail->unk4.seed)
-                    if (temp->checksum == checksum)
-                        return TRUE;
-    }
-
-    return FALSE;
-}
-
-u32 RestoreMailInfo(u8 *r0, u32 size)
-{
-    s32 index;
-    struct unkStruct_8094924 backup;
-    u32 temp;
-
-    xxx_init_struct_8094924_restore_809485C(&backup, r0, size);
-    for(index = 0; index < NUM_MAILBOX_SLOTS; index++)
-    {
-        RestoreWonderMail(&backup, &gUnknown_203B490->mailboxSlots[index]);
-    }
-    for(index = 0; index < MAX_ACCEPTED_JOBS; index++)
-    {
-        RestoreWonderMail(&backup, &gUnknown_203B490->pelipperBoardJobs[index]);
-    }
-    for(index = 0; index < MAX_ACCEPTED_JOBS; index++)
-    {
-        RestoreWonderMail(&backup, &gUnknown_203B490->jobSlots[index]);
-    }
-    for(index = 0; index < 0x38; index++)
-    {
-        RestoreIntegerBits(&backup, &temp, 1);
-        if(temp & 1)
-            gUnknown_203B490->PKMNNewsReceived[index] = TRUE;
-        else
-            gUnknown_203B490->PKMNNewsReceived[index] = FALSE;
-    }
-    RestoreIntegerBits(&backup, &temp, 1);
-    if(temp & 1)
-        gUnknown_203B490->unk328 = TRUE;
-    else
-        gUnknown_203B490->unk328 = FALSE;
-    
-    RestoreIntegerBits(&backup, gUnknown_203B490->unk190, 0x140);
-    RestoreIntegerBits(&backup, gUnknown_203B490->unk1B8, 0x3C0);
-    for(index = 0; index < 0x10; index++)
-    {
-        RestoreIntegerBits(&backup, &gUnknown_203B490->unk230[index].checksum, 0x20);
-        RestoreIntegerBits(&backup, &gUnknown_203B490->unk230[index].sub.seed, 0x18);
-        RestoreDungeonLocation(&backup, &gUnknown_203B490->unk230[index].sub.dungeon);
-    }
-    nullsub_102(&backup);
-    return backup.unk8;
-}
-
-u32 SaveMailInfo(u8 *r0, u32 size)
-{
-    s32 index;
-    struct unkStruct_8094924 backup;
-    u32 temp;
-
-    xxx_init_struct_8094924_save_809486C(&backup, r0, size);
-    for(index = 0; index < NUM_MAILBOX_SLOTS; index++)
-    {
-        SaveWonderMail(&backup, &gUnknown_203B490->mailboxSlots[index]);
-    }
-    for(index = 0; index < MAX_ACCEPTED_JOBS; index++)
-    {
-        SaveWonderMail(&backup, &gUnknown_203B490->pelipperBoardJobs[index]);
-    }
-    for(index = 0; index < MAX_ACCEPTED_JOBS; index++)
-    {
-        SaveWonderMail(&backup, &gUnknown_203B490->jobSlots[index]);
-    }
-    for(index = 0; index < 0x38; index++)
-    {
-        if(gUnknown_203B490->PKMNNewsReceived[index])
-            temp = -1;
-        else
-            temp = 0;
-        SaveIntegerBits(&backup, &temp, 1);
-    }
-    if(gUnknown_203B490->unk328)
-        temp = -1;
-    else
-        temp = 0;
-    SaveIntegerBits(&backup, &temp, 1);
-    SaveIntegerBits(&backup, gUnknown_203B490->unk190, 0x140);
-    SaveIntegerBits(&backup, gUnknown_203B490->unk1B8, 0x3C0);
-    for(index = 0; index < 0x10; index++)
-    {
-        SaveIntegerBits(&backup, &gUnknown_203B490->unk230[index].checksum, 0x20);
-        SaveIntegerBits(&backup, &gUnknown_203B490->unk230[index].sub.seed, 0x18);
-        SaveDungeonLocation(&backup, &gUnknown_203B490->unk230[index].sub.dungeon);
-    }
-    nullsub_102(&backup);
-    return backup.unk8;
-}
-
-void RestoreWonderMail(struct unkStruct_8094924 *a, WonderMail *b)
-{
-    RestoreIntegerBits(a, &b->mailType, 4);
-    RestoreIntegerBits(a, &b->missionType, 3);
-    RestoreIntegerBits(a, &b->unk2, 4);
-    RestoreIntegerBits(a, &b->clientSpecies, 9);
-    RestoreIntegerBits(a, &b->targetSpecies, 9);
-    RestoreIntegerBits(a, &b->targetItem, 8);
-    RestoreIntegerBits(a, &b->rewardType, 4);
-    RestoreIntegerBits(a, &b->itemReward, 8);
-    RestoreIntegerBits(a, &b->friendAreaReward, 6);
-    RestoreIntegerBits(a, &b->unk4.seed, 0x18);
-    RestoreDungeonLocation(a, &b->unk4.dungeon);
-}
-
-void SaveWonderMail(struct unkStruct_8094924 *a, WonderMail *b)
-{
-    SaveIntegerBits(a, &b->mailType, 4);
-    SaveIntegerBits(a, &b->missionType, 3);
-    SaveIntegerBits(a, &b->unk2, 4);
-    SaveIntegerBits(a, &b->clientSpecies, 9);
-    SaveIntegerBits(a, &b->targetSpecies, 9);
-    SaveIntegerBits(a, &b->targetItem, 8);
-    SaveIntegerBits(a, &b->rewardType, 4);
-    SaveIntegerBits(a, &b->itemReward, 8);
-    SaveIntegerBits(a, &b->friendAreaReward, 6);
-    SaveIntegerBits(a, &b->unk4.seed, 0x18);
-    SaveDungeonLocation(a, &b->unk4.dungeon);
-}
 
 void sub_80972F4(void)
 {
