+++ src/pelipper_board.c
@@ -14,11 +14,10 @@
 #include "pelipper_board.h"
 #include "code_80118A4.h"
 #include "wonder_mail_802C10C.h"
-#include "code_80958E8_1.h"
+#include "code_80958E8.h"
 #include "code_803B050.h"
 #include "wonder_mail_802C4C8.h"
 #include "wonder_mail_802C860.h"
-#include "code_8096AF8.h"
 
 EWRAM_DATA_2 struct unkStruct_203B308 *gPelipperBoard = {0};
 
@@ -79,9 +78,6 @@ extern void CreatePelipperBoardMenu(void);
 extern void sub_802EC10(void);
 
 extern void CreatePelipperAcceptedStatusBox(u32);
-
-extern void ResetPelipperBoardSlot(u8);
-extern void sub_80965F4(void);
 
 void SetPelipperBoardState(u32);
