+++ src/pokemon_news1.c
@@ -2,7 +2,7 @@
 #include "globaldata.h"
 #include "code_80118A4.h"
 #include "code_80130A8.h"
-#include "code_8096AF8.h"
+#include "code_80958E8.h"
 #include "constants/input.h"
 #include "input.h"
 #include "pokemon_mail.h"
