+++ src/pokemon_news3.c
@@ -3,7 +3,7 @@
 #include "code_800D090.h"
 #include "code_80118A4.h"
 #include "code_803B050.h"
-#include "code_80958E8_1.h"
+#include "code_80958E8.h"
 #include "constants/input.h"
 #include "constants/mailbox.h"
 #include "input.h"
