+++ src/post_office_guide2.c
@@ -3,7 +3,7 @@
 #include "code_800D090.h"
 #include "code_80118A4.h"
 #include "code_8094F88.h"
-#include "code_80958E8_1.h"
+#include "code_80958E8.h"
 #include "code_80972F4.h"
 #include "code_80A26CC.h"
 #include "constants/mailbox.h"
