+++ src/wonder_mail_802D098.c
@@ -3,8 +3,7 @@
 #include "code_80130A8.h"
 #include "code_802DE84.h"
 #include "code_803B050.h"
-#include "code_80958E8_1.h"
-#include "code_8096AF8.h"
+#include "code_80958E8.h"
 #include "code_80972F4.h"
 #include "code_8098BDC.h"
 #include "common_strings.h"
